+++ source3/lib/system_smbd.c
@@ -146,9 +146,9 @@ static int sys_getgrouplist(const char *user, gid_t gid, gid_t *groups, int *grp
 
 bool getgroups_unix_user(TALLOC_CTX *mem_ctx, const char *user,
 			 gid_t primary_gid,
-			 gid_t **ret_groups, size_t *p_ngroups)
+			 gid_t **ret_groups, uint32_t *p_ngroups)
 {
-	size_t ngrp;
+	uint32_t ngrp;
 	int max_grp;
 	gid_t *temp_groups;
 	gid_t *groups;
